Changes made on based various pr suggestions
