+++ Px.Utils.UnitTests/Validation/DatabaseValidation/DatabaseValidatorTests.cs
@@ -37,11 +37,20 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
         }
 
         [TestMethod]
-        public void ValidateDatabaseWithCustomValidatorsReturnsValidResult()
+        public void ValidateDatabaseWithInvalidDatabaseReturnsFeedback()
         {
             // Arrange
             MockFileSystem fileSystem = new();
             DatabaseValidator validator = new("database_invalid", fileSystem: fileSystem);
+            ValidationFeedbackKey invalidValueFormatKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.InvalidValueFormat);
+            ValidationFeedbackKey recommendedKeyMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.RecommendedKeyMissing);
+            ValidationFeedbackKey entryWithMultipleValuesKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.EntryWithMultipleValues);
+            ValidationFeedbackKey dataValidationFeedbackInvalidStructureKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidStructure);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowCountKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowCount);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowLengthKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowLength);
+            ValidationFeedbackKey unmatchingValueTypeKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.UnmatchingValueType);
+            ValidationFeedbackKey excessNewLinesInValueKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.ExcessNewLinesInValue);
+            ValidationFeedbackKey missingStubAndHeadingKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.MissingStubAndHeading);
 
             // Act
             ValidationResult result = validator.Validate();
@@ -50,14 +59,33 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
             Assert.IsNotNull(result, "Validation result should not be null");
             Assert.AreEqual(9, result.FeedbackItems.Count); // Unique feedbacks
             Assert.AreEqual(11, result.FeedbackItems.Values.SelectMany(f => f).Count()); // Total feedbacks including duplicates
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(invalidValueFormatKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(recommendedKeyMissingKey));
+            Assert.AreEqual(3, result.FeedbackItems[recommendedKeyMissingKey].Count); // 3 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(entryWithMultipleValuesKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidStructureKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowCountKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowLengthKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(unmatchingValueTypeKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(excessNewLinesInValueKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(missingStubAndHeadingKey));
         }
 
         [TestMethod]
-        public async Task ValidateDatabaseAsyncWithInvalidPxFileReturnsFeedback()
+        public async Task ValidateDatabaseAsyncWithInvaliDatabaseReturnsFeedback()
         {
             // Arrange
             MockFileSystem fileSystem = new();
             DatabaseValidator validator = new("database_invalid", fileSystem: fileSystem);
+            ValidationFeedbackKey invalidValueFormatKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.InvalidValueFormat);
+            ValidationFeedbackKey recommendedKeyMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.RecommendedKeyMissing);
+            ValidationFeedbackKey entryWithMultipleValuesKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.EntryWithMultipleValues);
+            ValidationFeedbackKey dataValidationFeedbackInvalidStructureKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidStructure);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowCountKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowCount);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowLengthKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowLength);
+            ValidationFeedbackKey unmatchingValueTypeKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.UnmatchingValueType);
+            ValidationFeedbackKey excessNewLinesInValueKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.ExcessNewLinesInValue);
+            ValidationFeedbackKey missingStubAndHeadingKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.MissingStubAndHeading);
 
             // Act
             ValidationResult result = await validator.ValidateAsync();
@@ -66,6 +94,16 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
             Assert.IsNotNull(result, "Validation result should not be null");
             Assert.AreEqual(9, result.FeedbackItems.Count); // Unique feedbacks
             Assert.AreEqual(11, result.FeedbackItems.Values.SelectMany(f => f).Count()); // Total feedbacks including duplicates
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(invalidValueFormatKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(recommendedKeyMissingKey));
+            Assert.AreEqual(3, result.FeedbackItems[recommendedKeyMissingKey].Count); // 3 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(entryWithMultipleValuesKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidStructureKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowCountKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowLengthKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(unmatchingValueTypeKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(excessNewLinesInValueKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(missingStubAndHeadingKey));
         }
 
         [TestMethod]
@@ -83,6 +121,16 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
                 customAliasFileValidators: customValidators,
                 customDirectoryValidators: customValidators,
                 fileSystem: fileSystem);
+            ValidationFeedbackKey invalidValueFormatKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.InvalidValueFormat);
+            ValidationFeedbackKey aliasFileMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.AliasFileMissing);
+            ValidationFeedbackKey recommendedKeyMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.RecommendedKeyMissing);
+            ValidationFeedbackKey entryWithMultipleValuesKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.EntryWithMultipleValues);
+            ValidationFeedbackKey dataValidationFeedbackInvalidStructureKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidStructure);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowCountKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowCount);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowLengthKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowLength);
+            ValidationFeedbackKey unmatchingValueTypeKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.UnmatchingValueType);
+            ValidationFeedbackKey excessNewLinesInValueKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.ExcessNewLinesInValue);
+            ValidationFeedbackKey missingStubAndHeadingKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.MissingStubAndHeading);
 
             // Act
             ValidationResult result = validator.Validate();
@@ -91,6 +139,18 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
             Assert.IsNotNull(result, "Validation result should not be null");
             Assert.AreEqual(10, result.FeedbackItems.Count); // Unique feedbacks
             Assert.AreEqual(26, result.FeedbackItems.Values.SelectMany(f => f).Count()); // Total feedbacks including duplicates
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(invalidValueFormatKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(aliasFileMissingKey));
+            Assert.AreEqual(15, result.FeedbackItems[aliasFileMissingKey].Count); // 15 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(recommendedKeyMissingKey));
+            Assert.AreEqual(3, result.FeedbackItems[recommendedKeyMissingKey].Count); // 3 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(entryWithMultipleValuesKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidStructureKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowCountKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowLengthKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(unmatchingValueTypeKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(excessNewLinesInValueKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(missingStubAndHeadingKey));
         }
 
         [TestMethod]
@@ -108,6 +168,17 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
                 customAliasFileValidators: customValidators,
                 customDirectoryValidators: customValidators,
                 fileSystem: fileSystem);
+            ValidationFeedbackKey invalidValueFormatKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.InvalidValueFormat);
+            ValidationFeedbackKey aliasFileMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.AliasFileMissing);
+            ValidationFeedbackKey recommendedKeyMissingKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.RecommendedKeyMissing);
+            ValidationFeedbackKey entryWithMultipleValuesKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.EntryWithMultipleValues);
+            ValidationFeedbackKey dataValidationFeedbackInvalidStructureKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidStructure);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowCountKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowCount);
+            ValidationFeedbackKey dataValidationFeedbackInvalidRowLengthKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.DataValidationFeedbackInvalidRowLength);
+            ValidationFeedbackKey unmatchingValueTypeKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.UnmatchingValueType);
+            ValidationFeedbackKey excessNewLinesInValueKey = new(ValidationFeedbackLevel.Warning, ValidationFeedbackRule.ExcessNewLinesInValue);
+            ValidationFeedbackKey missingStubAndHeadingKey = new(ValidationFeedbackLevel.Error, ValidationFeedbackRule.MissingStubAndHeading);
+
 
             // Act
             ValidationResult result = await validator.ValidateAsync();
@@ -116,6 +187,18 @@ namespace Px.Utils.UnitTests.Validation.DatabaseValidation
             Assert.IsNotNull(result, "Validation result should not be null");
             Assert.AreEqual(10, result.FeedbackItems.Count); // Unique feedbacks
             Assert.AreEqual(26, result.FeedbackItems.Values.SelectMany(f => f).Count()); // Total feedbacks including duplicates
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(invalidValueFormatKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(aliasFileMissingKey));
+            Assert.AreEqual(15, result.FeedbackItems[aliasFileMissingKey].Count); // 15 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(recommendedKeyMissingKey));
+            Assert.AreEqual(3, result.FeedbackItems[recommendedKeyMissingKey].Count); // 3 warnings
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(entryWithMultipleValuesKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidStructureKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowCountKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(dataValidationFeedbackInvalidRowLengthKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(unmatchingValueTypeKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(excessNewLinesInValueKey));
+            Assert.IsTrue(result.FeedbackItems.ContainsKey(missingStubAndHeadingKey));
         }
 
         [TestMethod]
