+++ Px.Utils.UnitTests/Validation/SyntaxValidationTests/StreamSyntaxValidationTests.cs
@@ -48,6 +48,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
 
             // Act
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
             Assert.AreEqual(8, result.Result.Count);
             Assert.AreEqual(0, feedback.Count);
         }
@@ -83,6 +84,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
 
             // Act
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
             Assert.AreEqual(10, result.Result.Count);
             Assert.AreEqual("YES", result.Result[8].Value);
             Assert.AreEqual("NO", result.Result[9].Value);
@@ -113,6 +115,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
 
             // Act
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
             Assert.AreEqual(4, result.FeedbackItems.Count);
             Assert.AreEqual(9, result.FeedbackItems[keyWhiteSpaceFeedbackKey][0].Line);
             Assert.AreEqual(12, result.FeedbackItems[valueWhiteSpaceFeedbackKey][0].Line);
@@ -494,6 +497,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
 
             // Act
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
             Assert.AreEqual(0, result.FeedbackItems.Count);
         }
 
@@ -513,6 +517,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
 
             // Act
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
             Assert.AreEqual(1, result.FeedbackItems.Count);
             Assert.AreEqual(9, result.FeedbackItems.First().Value[0].Line);
             Assert.AreEqual(16, result.FeedbackItems.First().Value[0].Character);
@@ -535,6 +540,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
             // Act
             SyntaxValidator validator = new(customValidationFunctions: new MockCustomSyntaxValidationFunctions());
             SyntaxValidationResult result = validator.Validate(stream, "foo", encoding);
+            stream.Close();
 
             // Assert
             Assert.AreEqual(0, result.FeedbackItems.Count);
@@ -553,6 +559,7 @@ namespace Px.Utils.UnitTests.SyntaxValidationTests
             // Act
             SyntaxValidator validator = new(customValidationFunctions: new MockCustomSyntaxValidationFunctions());
             SyntaxValidationResult result = await validator.ValidateAsync(stream, "foo", encoding);
+            stream.Close();
 
             // Assert
             Assert.AreEqual(0, result.FeedbackItems.Count);
