+++ Px.Utils/Validation/ContentValidation/ContentValidator.cs
@@ -1,4 +1,5 @@
 ﻿using Px.Utils.PxFile;
+using Px.Utils.Validation.DatabaseValidation;
 using Px.Utils.Validation.SyntaxValidation;
 using System.Text;
 
@@ -14,13 +15,13 @@ namespace Px.Utils.Validation.ContentValidation
     /// </summary>
     public sealed partial class ContentValidator(
         string filename,
-        Encoding? encoding,
+        Encoding encoding,
         ValidationStructuredEntry[] entries,
         CustomContentValidationFunctions? customContentValidationFunctions = null,
         PxFileSyntaxConf? syntaxConf = null) : IValidator
     {
         private readonly string _filename = filename;
-        private readonly Encoding? _encoding = encoding;
+        private readonly Encoding _encoding = encoding;
         /// <summary>
         /// Default language of the Px file defined with LANGUAGE keyword by default
         /// </summary>
