+++ Px.Utils/Validation/IPxFileStreamValidator.cs
@@ -15,9 +15,8 @@ namespace Px.Utils.Validation
         /// <param name="filename">Name of the PX file</param>
         /// <param name="encoding">Encoding of the PX file. If not provided, validator tries to find the encoding.</param>
         /// <param name="fileSystem">File system to use for file operations. If not provided, default file system is used.</param>
-        /// <param name="leaveStreamOpen">Boolean value that determines whether the stream should be left open after validation.
         /// <returns><see cref="ValidationResult"/> object that contains an array of feedback items gathered during the validation process.</returns>
-        public ValidationResult Validate(Stream stream, string filename, Encoding? encoding = null, bool leaveStreamOpen = false, IFileSystem? fileSystem = null);
+        public ValidationResult Validate(Stream stream, string filename, Encoding? encoding = null, IFileSystem? fileSystem = null);
     }
 
     /// <summary>
@@ -31,10 +30,9 @@ namespace Px.Utils.Validation
         /// <param name="stream">Stream of the PX file to be validated</param>
         /// <param name="filename">Name of the PX file</param>
         /// <param name="encoding">Encoding of the PX file. If not provided, validator tries to find the encoding.</param>
-        /// <param name="leaveStreamOpen">Boolean value that determines whether the stream should be left open after validation.
         /// <param name="fileSystem">File system to use for file operations. If not provided, default file system is used.</param>
         /// <param name="cancellationToken">Cancellation token for cancelling the validation process</param>
         /// <returns><see cref="ValidationResult"/> object that contains an array of feedback items gathered during the validation process.</returns>
-        public Task<ValidationResult> ValidateAsync(Stream stream, string filename, Encoding? encoding = null, bool leaveStreamOpen = false, IFileSystem? fileSystem = null, CancellationToken cancellationToken = default);
+        public Task<ValidationResult> ValidateAsync(Stream stream, string filename, Encoding? encoding = null, IFileSystem? fileSystem = null, CancellationToken cancellationToken = default);
     }
 }
